Tidied up some code and renamed some folders
