+++ src/lib/offscreen-worker.tsx
@@ -4,11 +4,10 @@ import * as Comlink from "comlink";
 import * as THREE from "three";
 import { extend, createRoot, events } from "@react-three/fiber";
 import { Scene } from "../components/Scene";
-import { initTransferHandler } from "../utils/event.transferhandler";
+import { initTransferHandler } from "../lib/event.transferhandler";
 import { observable } from "@legendapp/state";
 
 initTransferHandler();
-
 extend(THREE);
 
 let root;
